backend/requirements.txt
anyjson>=0.3.3
argcomplete>=0.8.7
astroid>=1.4.5
celery>=3.1.23
celerymon>=1.0.3
certifi>=14.5.14
colorama>=0.3.7
coverage>=4.1
defusedxml>=0.4.1
Django>=2.0.0
django-admin-bootstrapped>=2.3.2
django-annoying>=0.8.0
django-appconf>=0.6
django-autocomplete-light>=3.1.4
django-bootstrap3>=7.0.1
django-celery>=3.1.16
django-compressor>=1.4
django-cors-headers>=1.1.0
django-crispy-forms>=1.4.0
django-debug-toolbar>=1.2.2
django-devserver>=0.8.0
django-extensions>=1.6.1
django-extra-views>=0.6.4
django-filter>=0.9.2
django-localflavor>=1.1
django-model-utils>=2.2
django-tastypie>=0.13.3
django-tastypie-swagger>=0.1.4
djangorestframework>=3.3.3
Jinja2>=2.8
psycopg2>=2.8.4
redis>=2.10.3
requests>=2.5.1
requests-oauthlib>=0.4.2
simplejson>=3.6.5
six>=1.9.0
mysqlclient>=1.4.6
selenium
